Reworked banner, added footer, added black lettering logo
I reworked the banner so that it looks kind of decent, but feel free to
change it however you'd like. I also added a footer at the bottom with a
contact link that sends a message to my islander email with a subject
line that says "You guys get an A". I also added a black lettered
version of the logo for the venue.